lower pairwise operations with the same input key into an element wise operation
