+++ fusor-ml/core/src/composite/mod.rs
@@ -7,3 +7,4 @@ mod softmax;
 mod unsqueeze;
 pub use unsqueeze::Unsqueeze;
 mod rope;
+mod sqr;
